move things  because none of it is ready to be shared yet
